repair compressed mode
 1. Extend file pointers from unsigned int to unsigned long to make it
    64-bit compatible on 64-bit hosts.
 2. Repair linalloc() overflow detection.
 3. Increased linalloc() memory from 12KB to 32KB.
 4. Decreased frequency of show_progress() updates (slow serial console)

Unfortunately this doesn't help me on aarch64 at all as the caches are
disabled so uncompressing modules takes ages :-/

+++ server/src/gunzip.h
@@ -18,9 +18,9 @@ l4_addr_t gunzip_upper_mem_linalloc(void);
 #define RAW_ADDR(x) (x)
 #define RAW_SEG(x) (x)
 
-extern unsigned int filepos;
-extern unsigned int filemax;
-extern unsigned int fsmax;     /* max size of fs/readable media */
+extern unsigned long filepos;
+extern unsigned long filemax;
+extern unsigned long fsmax;     /* max size of fs/readable media */
 extern unsigned int compressed_file;
 extern grub_error_t errnum;
